requirements.txt
Flask==3.1.0
google-auth==2.38.0
google-genai==1.2.0
annotated-types==0.7.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
idna==3.10
Jinja2==3.1.5
MarkupSafe==3.0.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
requests==2.32.3
rsa==4.9
urllib3==2.3.0
websockets==14.2
Werkzeug==3.1.3